got tests to pass, but still not sure if it's right
